format: introduce `-format table|yaml`, use `table` by default

+++ main.go
@@ -2,27 +2,55 @@ package main
 
 import (
 	"context"
+	"flag"
 	"fmt"
+	"io"
 	"net/http"
 	"os"
 	"os/signal"
 	"time"
 
 	"github.com/mxmCherry/taxes/v2/internal/bankgovua"
+	"github.com/mxmCherry/taxes/v2/internal/format"
 	"github.com/mxmCherry/taxes/v2/internal/tax"
 	"golang.org/x/time/rate"
 	"gopkg.in/yaml.v3"
+
+	ftable "github.com/mxmCherry/taxes/v2/internal/format/table"
+	fyaml "github.com/mxmCherry/taxes/v2/internal/format/yaml"
 )
 
+var formats map[string]func(io.Writer) format.Formatter
+
+var flags struct {
+	Format string
+}
+
+func init() {
+	formats = map[string]func(io.Writer) format.Formatter{
+		"yaml":  fyaml.New,
+		"table": ftable.New,
+	}
+	flag.StringVar(&flags.Format, "format", "table", "output format: table|yaml")
+}
+
 func main() {
-	output := yaml.NewEncoder(os.Stdout)
-	defer output.Close()
+	flag.Parse()
+
+	newFormatter, ok := formats[flags.Format]
+	if !ok {
+		fmt.Fprintf(os.Stderr, "ERROR: unsupported -format %q", flags.Format)
+		os.Exit(1)
+	}
+
+	formatter := newFormatter(os.Stdout)
+	defer formatter.Close()
 
 	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, os.Kill)
 	defer cancel()
 
-	for _, filename := range os.Args[1:] {
-		if err := process(ctx, output, filename); err != nil {
+	for _, filename := range flag.Args() {
+		if err := process(ctx, formatter, filename); err != nil {
 			fmt.Fprintf(os.Stderr, "ERROR: %q: %s\n", filename, err)
 			return
 		}
@@ -34,7 +62,7 @@ func main() {
 	}
 }
 
-func process(ctx context.Context, output *yaml.Encoder, filename string) error {
+func process(ctx context.Context, formatter format.Formatter, filename string) error {
 	f, err := os.Open(filename)
 	if err != nil {
 		return fmt.Errorf("open %q: %w", filename, err)
@@ -58,7 +86,7 @@ func process(ctx context.Context, output *yaml.Encoder, filename string) error {
 		return fmt.Errorf("tax calc: %w", err)
 	}
 
-	if err := output.Encode(run); err != nil {
+	if err := formatter.Format(&run); err != nil {
 		return fmt.Errorf("output yaml: %w", err)
 	}
 	return nil
